// src/shared/api/generated/models/supportCaseSearchQueryDtoAssignmentStatesItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoAssignmentStatesItem =
  (typeof SupportCaseSearchQueryDtoAssignmentStatesItem)[keyof typeof SupportCaseSearchQueryDtoAssignmentStatesItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoAssignmentStatesItem = {
  ASSIGNED: "ASSIGNED",
  UNASSIGNED: "UNASSIGNED",
} as const;