// src/shared/api/generated/models/supportCaseSearchQueryDtoCategoryCodesItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoCategoryCodesItem =
  (typeof SupportCaseSearchQueryDtoCategoryCodesItem)[keyof typeof SupportCaseSearchQueryDtoCategoryCodesItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoCategoryCodesItem = {
  INFORMATION_REQUEST: "INFORMATION_REQUEST",
  PROBLEM_RESOLUTION: "PROBLEM_RESOLUTION",
  DECISION_SUPPORT: "DECISION_SUPPORT",
  ACTION_REQUEST: "ACTION_REQUEST",
  FEEDBACK: "FEEDBACK",
  OTHER: "OTHER",
} as const;