// src/shared/api/generated/models/supportCaseSearchQueryDtoChannelsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoChannelsItem =
  (typeof SupportCaseSearchQueryDtoChannelsItem)[keyof typeof SupportCaseSearchQueryDtoChannelsItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoChannelsItem = {
  TEXT: "TEXT",
  VOICE: "VOICE",
  CMS: "CMS",
} as const;