// src/shared/api/generated/models/supportCaseSearchQueryDtoDeliveryState.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoDeliveryState =
  (typeof SupportCaseSearchQueryDtoDeliveryState)[keyof typeof SupportCaseSearchQueryDtoDeliveryState];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoDeliveryState = {
  PROBLEM: "PROBLEM",
  HEALTHY: "HEALTHY",
} as const;