// src/shared/api/generated/models/supportCaseSearchQueryDtoDraftState.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoDraftState =
  (typeof SupportCaseSearchQueryDtoDraftState)[keyof typeof SupportCaseSearchQueryDtoDraftState];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoDraftState = {
  HAS_DRAFT: "HAS_DRAFT",
  NO_DRAFT: "NO_DRAFT",
} as const;