// src/shared/api/generated/models/supportCaseSearchQueryDtoPrioritiesItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoPrioritiesItem =
  (typeof SupportCaseSearchQueryDtoPrioritiesItem)[keyof typeof SupportCaseSearchQueryDtoPrioritiesItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoPrioritiesItem = {
  LOW: "LOW",
  NORMAL: "NORMAL",
  HIGH: "HIGH",
  URGENT: "URGENT",
  CRITICAL: "CRITICAL",
} as const;