// src/shared/api/generated/models/supportCaseSearchQueryDtoSlaStatesItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoSlaStatesItem =
  (typeof SupportCaseSearchQueryDtoSlaStatesItem)[keyof typeof SupportCaseSearchQueryDtoSlaStatesItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoSlaStatesItem = {
  ON_TRACK: "ON_TRACK",
  AT_RISK: "AT_RISK",
  BREACHED: "BREACHED",
  NOT_CONFIGURED: "NOT_CONFIGURED",
} as const;