// src/shared/api/generated/models/supportCaseSearchQueryDtoStatusesItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportCaseSearchQueryDtoStatusesItem =
  (typeof SupportCaseSearchQueryDtoStatusesItem)[keyof typeof SupportCaseSearchQueryDtoStatusesItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportCaseSearchQueryDtoStatusesItem = {
  OPEN: "OPEN",
  IN_PROGRESS: "IN_PROGRESS",
  WAITING_END_USER: "WAITING_END_USER",
  WAITING_SYSTEM: "WAITING_SYSTEM",
  WAITING_ADMIN: "WAITING_ADMIN",
  RESOLVED: "RESOLVED",
  UNRESOLVED: "UNRESOLVED",
  CANCELLED: "CANCELLED",
} as const;